Time: 20 ms (100.00%), Space: 5.9 MB (82.61%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -3,7 +3,7 @@ public:
     int concatenatedBinary(int n) {
         long result = 1L, numberOfBits = 1;
         
-        for (int i = 2; i <= n; i++) {
+        for (long i = 2; i <= n; i++) {
             if (i == (1 << numberOfBits)) {
                 numberOfBits++;
             }
